- download_images 와 perform resize thread로 동시에 처리 - que에 다운로드 받은 파일 저장하면 perform_thread가  Que에서 읽어감

+++ thumbnail_maker.py
@@ -4,64 +4,36 @@ import os
 import logging
 from urllib.parse import urlparse
 from urllib.request import urlretrieve
-import threading
+import threading, queue
 import PIL
 from PIL import Image
-
-FORMAT = "[%(threadName)s , %(asctime)s, %(levelname)s, %(message)s]"
-logging.basicConfig(filename='logfile.log', level=logging.DEBUG, format=FORMAT)
-
-
-# Purpose: thread worker에게 다운받은 이미지 사이지를 인스턴스변수 downloaded_byte에 누적저장
+logging.basicConfig(level=logging.DEBUG)
+#logging.basicConfig(filename='logfile.log', level=logging.DEBUG)
 
 class ThumbnailMakerService(object):
     def __init__(self, home_dir='.'):
         self.home_dir = home_dir
         self.input_dir = self.home_dir + os.path.sep + 'incoming'
         self.output_dir = self.home_dir + os.path.sep + 'outgoing'
-
-        # 1.threading.Lock() 개체 생성
-        self.dl_lock = threading.Lock()
-        self.downloaded_bytes = 0
-        max_concurrent_dl = 4
-        self.dl_sem = threading.Semaphore(max_concurrent_dl)
-
-
-    def download_image(self, url):
-        with self.dl_sem:
-            img_filename = urlparse(url).path.split('/')[-1]
-            dest_path = self.input_dir + os.path.sep + img_filename
-            urlretrieve(url, dest_path)
-            img_size = os.path.getsize(dest_path)
-
-            with self.dl_lock:
-                self.downloaded_bytes += img_size
-            logging.info(f"image {img_size} saved to {dest_path}")
+        self.que = queue.Queue()
 
     def download_images(self, img_url_list):
-        # validate inputs
-        if not img_url_list:
-            return
+
         os.makedirs(self.input_dir, exist_ok=True)
         logging.info("beginning image downloads")
 
         start = time.perf_counter()
-        #---- Stat- 쓰레드 작업으로 돌릴 부분
-        # for url in img_url_list:
-        #     img_filename = urlparse(url).path.split('/')[-1]
-        #     urlretrieve(url, self.input_dir + os.path.sep + img_filename)
-        #---- End - 쓰레드 작업으로 돌릴 부분
-
-        threads = []
         for url in img_url_list:
-            t =  threading.Thread(target=self.download_image,args=(url,))
-            t.start()
-            threads.append(t)
-
-        for _t in threads:
-            _t.join()
+            # download each image and save to the input dir 
+            img_filename = urlparse(url).path.split('/')[-1]
+            dest_path = self.input_dir + os.path.sep + img_filename
+            urlretrieve(url, dest_path)
+            self.que.put(img_filename)
+            logging.info(f"- downloaded: {img_filename}")
         end = time.perf_counter()
-        logging.info(f"downloaded {len(img_url_list)} images  in {end-start} seconds: {self.downloaded_bytes} TOTAL BYTES")
+
+        self.que.put(None)
+        logging.info("downloaded {} images in {} seconds".format(len(img_url_list), end - start))
 
     def perform_resizing(self):
         # validate inputs
@@ -74,33 +46,46 @@ class ThumbnailMakerService(object):
         num_images = len(os.listdir(self.input_dir))
 
         start = time.perf_counter()
-        for filename in os.listdir(self.input_dir):
-            orig_img = Image.open(self.input_dir + os.path.sep + filename)
-            for basewidth in target_sizes:
-                img = orig_img
-                # calculate target height of the resized image to maintain the aspect ratio
-                wpercent = (basewidth / float(img.size[0]))
-                hsize = int((float(img.size[1]) * float(wpercent)))
-                # perform resizing
-                img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
-                
-                # save the resized image to the output dir with a modified file name 
-                new_filename = os.path.splitext(filename)[0] + \
-                    '_' + str(basewidth) + os.path.splitext(filename)[1]
-                img.save(self.output_dir + os.path.sep + new_filename)
+        while True:
+            filename =  self.que.get()
+            if filename:
+                logging.info(f"resizing image :{filename}")
+                orig_img = Image.open(self.input_dir + os.path.sep + filename)
+                for basewidth in target_sizes:
+                    img = orig_img
+                    # calculate target height of the resized image to maintain the aspect ratio
+                    wpercent = (basewidth / float(img.size[0]))
+                    hsize = int((float(img.size[1]) * float(wpercent)))
+                    # perform resizing
+                    img = img.resize((basewidth, hsize), PIL.Image.LANCZOS)
 
-            os.remove(self.input_dir + os.path.sep + filename)
+                    # save the resized image to the output dir with a modified file name 
+                    new_filename = os.path.splitext(filename)[0] + \
+                        '_' + str(basewidth) + os.path.splitext(filename)[1]
+                    img.save(self.output_dir + os.path.sep + new_filename)
+
+                os.remove(self.input_dir + os.path.sep + filename)
+                logging.info(f"done resizing: {self.input_dir + os.path.sep + new_filename}")
+                self.que.task_done()
+
+            else:
+                self.que.task_done()
+                break
+
         end = time.perf_counter()
-
         logging.info("created {} thumbnails in {} seconds".format(num_images, end - start))
-
+        
     def make_thumbnails(self, img_url_list):
         logging.info("START make_thumbnails")
         start = time.perf_counter()
 
-        self.download_images(img_url_list)
-#        self.perform_resizing()
+        t1 = threading.Thread(target=self.download_images, args=([img_url_list]))
+        t2 = threading.Thread(target=self.perform_resizing)
 
+        t1.start()
+        t2.start()
+        t1.join()
+        t2.join()
         end = time.perf_counter()
         logging.info("END make_thumbnails in {} seconds".format(end - start))
